feature: add support for navigation data (127237, 129283, 129284)
Add support for emitting the various PGNs that carry the information
about the distance and bearing to the active waypoint. This should
provide all information needed by autopilots (heading to steer, XTE)
as well as bearing/distance to waypoint and estimated time of arrival,
which is commonly displayed for human consumption.

Tested with OpenCPN as the producer of APB sentences, and a B&G Triton2
display to show the various data items.

This is based on the outdated and incomplete implementation that lives
on the 'autopilot' branch.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>